+++ angular-workspace/projects/example-client-app/src/app/app.component.ts
@@ -1,8 +1,11 @@
 import { Component } from '@angular/core';
+import { NimbleTheme } from '@ni/nimble-components/dist/esm/theme-provider/themes';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss']
 })
-export class AppComponent { }
+export class AppComponent {
+    public theme: NimbleTheme = NimbleTheme.Light;
+}