+++ packages/nimble-components/src/number-field/styles.ts
@@ -1,8 +1,11 @@
 import { css } from '@microsoft/fast-element';
+import { display } from '@microsoft/fast-foundation';
 import {
     borderColor,
+    borderColorRgb,
     borderColorHover,
     borderWidth,
+    contentFontColor,
     contentFontColorDisabled,
     contentFontSize,
     controlHeight,
@@ -12,19 +15,21 @@ import {
     labelFontColor,
     labelFontFamily,
     labelFontSize,
+    labelFontWeight,
     labelHeight,
     labelTextTransform,
     smallDelay
 } from '../theme-provider/design-tokens';
 
 export const styles = css`
+    ${display('inline-block')}
+
     :host {
-        display: inline-block;
         font-family: ${fontFamily};
         font-size: ${contentFontSize};
         outline: none;
         user-select: none;
-        color: ${labelFontColor};
+        color: ${contentFontColor};
         height: calc(${labelHeight} + ${controlHeight});
     }
 
@@ -40,7 +45,7 @@ export const styles = css`
         flex-direction: row;
         border-radius: 0px;
         font-family: ${fontFamily};
-        border-bottom: ${borderWidth} solid ${borderColor};
+        border-bottom: ${borderWidth} solid rgba(${borderColorRgb}, 0.3);
         padding-bottom: 1px;
         transition: border-bottom ${smallDelay}, padding-bottom ${smallDelay};
     }
@@ -88,7 +93,6 @@ export const styles = css`
 
     .control::placeholder {
         color: ${labelFontColor};
-        opacity: 0.6;
     }
 
     .control:focus-within::placeholder {
@@ -101,8 +105,10 @@ export const styles = css`
 
     .label {
         display: flex;
+        color: ${labelFontColor};
         font-family: ${labelFontFamily};
         font-size: ${labelFontSize};
+        font-weight: ${labelFontWeight};
         line-height: ${labelHeight};
         text-transform: ${labelTextTransform};
     }