+++ packages/nimble-components/src/number-field/tests/number-field.stories.ts
@@ -16,6 +16,12 @@ const metadata: Meta<NumberFieldArgs> = {
     title: 'Number Field',
     decorators: [withXD],
     parameters: {
+        docs: {
+            description: {
+                component:
+                    'Similar to a single line text box but only used for numeric data. The controls allow the user to increment and decrement the value.'
+            }
+        },
         design: {
             artboardUrl:
                 'https://xd.adobe.com/view/8ce280ab-1559-4961-945c-182955c7780b-d9b1/screen/eaa9ee19-4411-4648-b19d-41f61f9a01cf/specs/'