+++ packages/nimble-tokens/.eslintrc.js
@@ -3,6 +3,18 @@ module.exports = {
     extends: [
         '@ni/eslint-config-javascript'
     ],
-    rules: {
-    }
+    overrides: [
+        {
+            files: [
+                'dist/styledictionary/nimble-extensions.js'
+            ],
+            rules: {
+                // Okay to use dev dependencies in build scripts
+                'import/no-extraneous-dependencies': 'off',
+
+                // Okay to use console.log in build scripts
+                'no-console': 'off'
+            }
+        }
+    ]
 };